requirements.txt
streamlit==1.16.0
faiss-cpu==1.7.2
sentence-transformers==2.2.0
langchain==0.0.204
langchain-ollama==0.1.0
numpy==1.24.2
torch==2.0.0
pydantic==1.10.2
pyyaml==6.0
